+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * [API Change] When loading connections from a `.ini` file via `%sql --section section_name`, the section name is set as the connection alias
 * [API Change] Starting connections from a `.ini` file via `%sql [section_name]` has been deprecated
 * [Doc] Fixes documentation inaccuracy that said `:variable` was deprecated (we brought it back in `0.9.0`)
+* [Fix] Fix an error that caused a connection error message to be turned into a `print` statement
 * [Fix] Fix Twice message printing when switching to the current connection (#772)
 * [Fix] Error when using %sqlplot in snowflake (#697)